+++ src/main/java/com/appcenter/marketplace/domain/coupon/CouponRepository.java
@@ -1,6 +1,5 @@
-package com.appcenter.marketplace.domain.coupon.repository;
+package com.appcenter.marketplace.domain.coupon;
 
-import com.appcenter.marketplace.domain.coupon.domain.Coupon;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
